+++ definitions/base.py
@@ -11,11 +11,14 @@ __revision__ = '$Format:%H$'
 
 @unique
 class InputType(Enum):
-    Layer = 'Layer'
-    Field = 'Field'
+    Color = 'Color'
     CheckBox = 'CheckBox'
-    SpinBox = 'SpinBox'
+    Field = 'Field'
+    Fields = 'Fields'
+    Layer = 'Layer'
     List = 'List'
+    SpinBox = 'SpinBox'
+    Text = 'Text'
 
 
 class BaseDefinitions: